requirements.txt
django
django-import-export
django-crispy-forms
python-decouple
arrow
requests
openpyxl
pyodbc
pandas
